+++ src/hardware.rs
@@ -9,7 +9,7 @@ pub mod pot;
 pub mod sinkuart;
 
 pub trait Hardware {
-    fn update(&mut self, time: usize);
-    fn event(&mut self, _time: usize, _event: &String) { }
+    fn update(&mut self, time: u64);
+    fn event(&mut self, _time: u64, _event: &String) { }
     //fn get_pin(&self, name: String) -> Weak<RefCell<PinState>>;
 }
